Safe Commit, história finalizando de página de exibição dos pratos montáveis e personalização dos pratos

+++ db/migrate/20161016045423_create_alimento_precos.rb
@@ -2,7 +2,7 @@ class CreateAlimentoPrecos < ActiveRecord::Migration
   def change
     create_table :alimento_precos do |t|
       t.string :nome
-      t.integer :preco
+      t.float :preco
       t.references :prato_montado, index: true, foreign_key: true
 
       t.timestamps null: false
